+++ src/robots/sixi.cpp
@@ -99,7 +99,7 @@ void IK(const float *const axies, int32_t *motorStepArray) {
    @param axies the resulting cartesian coordinate
    @return 0 if no problem, 1 on failure.
 */
-int FK(long *motorStepArray, float *axies) {
+int FK(uint32_t *motorStepArray, float *axies) {
   // TODO fill me in!
 
   return 0;
