+++ src/robots/stewartLinear.cpp
@@ -95,7 +95,7 @@ void IK(const float *const cartesian, int32_t *motorStepArray) {
    @param axies the resulting cartesian coordinate
    @return 0 if no problem, 1 on failure.
 */
-int FK(long *motorStepArray, float *axies) {
+int FK(uint32_t *motorStepArray, float *axies) {
   return 0;
 }
